feat: add a function to control the track width
the function named check_trackwidth has the similar task as che_radius, but this time control the track width. User can also set the minimum trackwidth and the tolerant cones. It returns not only the new adding cones which fulfill the condition, it also returns the relevant position of every cone to the track on the other side, which will be usefull for the further function of finding the middle line of the track.